+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## 1.1.0
 
 - Add flushIp() service method.
-- Add setFailCountSeconds() config method.
 - Auto-expire entries in failCount table.
+- Add setFailCountSeconds() config method.
+- Auto-expire entries in blockCount table.
+- Add setBlockCountSeconds() config method.
 
 ## 1.0.0